+++ gidgethub/sansio.py
@@ -386,11 +386,12 @@ def format_url(url: str, url_vars: Mapping[str, Any], *, base_url: str = DOMAIN)
     return expanded_url
 
 
-def get_jwt(*, app_id: str, private_key: str) -> str:
-    """Construct the JWT token, used for GitHub App authentication."""
+def _get_jwt(*, app_id: str, private_key: str) -> str:
+    """Construct the JWT (JSON Web Token), used for GitHub App authentication."""
+    time_int = int(time.time())
     payload = {
-        "iat": int(time.time()),
-        "exp": int(time.time()) + (10 * 60),
+        "iat": time_int,
+        "exp": time_int + (10 * 60),
         "iss": app_id,
     }
     encoded = jwt.encode(payload, private_key, algorithm="RS256")
@@ -410,7 +411,7 @@ async def get_installation_access_token(
     Doc: https://developer.github.com/v3/apps/#create-a-new-installation-token
     """
     access_token_url = f"/app/installations/{installation_id}/access_tokens"
-    jwt = get_jwt(app_id=app_id, private_key=private_key)
+    jwt = _get_jwt(app_id=app_id, private_key=private_key)
     response = await gh.post(
         access_token_url,
         data=b"",