+++ tests/test_sansio.py
@@ -2,9 +2,12 @@ import datetime
 import http
 import json
 import pathlib
+from unittest import mock
 
-
+import jwt
 import pytest
+
+from .test_abc import MockGitHubAPI
 
 from gidgethub import (
     BadRequest,
@@ -419,7 +422,7 @@ class TestDecipherResponse:
             "link": '<http://example.com>; test="unimportant"',
         }
         data = {
-            "id": 208_045_946,
+            "id": 208045946,
             "url": "https://api.github.com/repos/octocat/Hello-World/labels/bug",
             "name": "bug",
             "color": "f29513",
@@ -505,7 +508,7 @@ class TestDecipherResponse:
             "link": '<http://example.com>; test="unimportant"',
         }
         data = {
-            "id": 208_045_946,
+            "id": 208045946,
             "url": "https://api.github.com/repos/octocat/Hello-World/labels/bug",
             "name": "bug",
             "color": "f29513",
@@ -587,3 +590,44 @@ class TestFormatUrl:
         label = {"name": "CLA signed"}
         url = sansio.format_url(template_url, label, base_url=base_url)
         assert url == "https://api.github.com/repos/python/cpython/labels/CLA%20signed"
+
+
+class TestGitHubAppUtils:
+
+    """Tests for GitHub App utilities."""
+
+    @mock.patch("time.time")
+    def test_get_jwt(self, time_mock):
+        app_id = 12345
+
+        # copied from https://github.com/jpadilla/pyjwt/blob/master/tests/keys/testkey_rsa
+        time_mock.return_value = 1587069751.5588422
+
+        with open("tests/samples/test_rsa_key", "r") as fp:
+            private_key = fp.read()
+
+            result = sansio._get_jwt(app_id=app_id, private_key=private_key)
+            expected_payload = {
+                "iat": 1587069751,
+                "exp": 1587069751 + (10 * 60),
+                "iss": app_id,
+            }
+
+            assert result == jwt.encode(
+                expected_payload, private_key, algorithm="RS256"
+            ).decode("utf-8")
+
+    @pytest.mark.asyncio
+    async def test_get_installation_access_token(self):
+        gh = MockGitHubAPI()
+        installation_id = 6789
+        app_id = 12345
+
+        with open("tests/samples/test_rsa_key", "r") as fp:
+            private_key = fp.read()
+
+        await sansio.get_installation_access_token(
+            gh, installation_id=installation_id, app_id=app_id, private_key=private_key
+        )
+
+        assert gh.url == "https://api.github.com/app/installations/6789/access_tokens"