+++ docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+4.1.0
+'''''
+
+.. note::
+   Under development
+
+- Add :meth:`gidgethub.sansio.get_installation_access_token` for obtaining an
+  installation access token that can be used to authenticate as a GitHub App
+  installation. (`Issue #71 <https://github.com/brettcannon/gidgethub/issues/71>`_
+
 4.0.0
 '''''